requirements.txt
environs==9.5.0
python-telegram-bot==11.1.0
redis==4.4.0
requests==2.28.1

